[Behat] Extract Country context
