+++ etc/behat/suites/ui/addressing.yml
@@ -7,7 +7,7 @@ default:
             contexts_as_services:
                 - sylius.behat.context.hook.doctrine_orm
 
-                - sylius.behat.context.transform.addressing
+                - sylius.behat.context.transform.country
 
                 - sylius.behat.context.setup.security
