+++ src/Sylius/Behat/spec/Context/Transform/AddressingContextSpec.php
@@ -15,9 +15,7 @@ use Behat\Behat\Context\Context;
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Addressing\Converter\CountryNameConverterInterface;
 use Sylius\Component\Addressing\Model\AddressInterface;
-use Sylius\Component\Addressing\Model\CountryInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
-use Sylius\Component\Resource\Repository\RepositoryInterface;
 
 /**
  * @author Łukasz Chruściel <lukasz.chrusciel@lakion.com>
@@ -26,10 +24,9 @@ class AddressingContextSpec extends ObjectBehavior
 {
     function let(
         FactoryInterface $addressFactory,
-        CountryNameConverterInterface $countryNameConverter,
-        RepositoryInterface $countryRepository
+        CountryNameConverterInterface $countryNameConverter
     ) {
-        $this->beConstructedWith($addressFactory, $countryNameConverter, $countryRepository);
+        $this->beConstructedWith($addressFactory, $countryNameConverter);
     }
 
     function it_is_initializable()
@@ -59,34 +56,5 @@ class AddressingContextSpec extends ObjectBehavior
         $address->setPostcode('90210')->shouldBeCalled();
 
         $this->createNewAddress('France')->shouldReturn($address);
-    }
-
-    function it_gets_country_based_on_given_country_name(
-        CountryNameConverterInterface $countryNameConverter,
-        RepositoryInterface $countryRepository,
-        CountryInterface $country
-    ) {
-        $countryNameConverter->convertToCode('France')->willReturn('FR');
-        $countryRepository->findOneBy(['code' => 'FR'])->willReturn($country);
-        $country->getCode()->willReturn('FR');
-
-        $this->getCountryByName('France')->shouldReturn($country);
-    }
-
-    function it_throws_invalid_argument_exception_when_country_is_missing(
-        CountryNameConverterInterface $countryNameConverter,
-        RepositoryInterface $countryRepository
-    ) {
-        $countryNameConverter->convertToCode('France')->willReturn('FR');
-        $countryRepository->findOneBy(['code' => 'FR'])->willReturn(null);
-
-        $this->shouldThrow(\InvalidArgumentException::class)->during('getCountryByName', ['France']);
-    }
-
-    function it_throws_invalid_argument_exception_when_cannot_convert_name_to_code(CountryNameConverterInterface $countryNameConverter)
-    {
-        $countryNameConverter->convertToCode('France')->willThrow(\InvalidArgumentException::class);
-
-        $this->shouldThrow(\InvalidArgumentException::class)->during('getCountryByName', ['France']);
     }
 }
